requirements.txt
# AI Video Generator Requirements

# Core PyTorch
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# AI Models
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0

# Video Processing
moviepy>=1.0.3
opencv-python>=4.8.0
pillow>=9.5.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# System Monitoring
psutil>=5.9.0
GPUtil>=1.4.0

# GUI (Optional)
customtkinter>=5.2.0
tkinter

# Utilities
numpy>=1.24.0
scipy>=1.10.0
tqdm>=4.65.0
requests>=2.31.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional GPU monitoring
nvidia-ml-py>=11.0.0

# File handling
pathlib2>=2.3.7
click>=8.1.0

# Configuration
pyyaml>=6.0
configparser>=5.3.0

# Logging
loguru>=0.7.0

# Progress bars
rich>=13.0.0

# Memory profiling (optional)
memory-profiler>=0.60.0

# Performance monitoring
py-spy>=0.3.14
